+++ app/build.gradle
@@ -46,8 +46,6 @@ android {
         targetSdkVersion 33
         versionCode 1
         versionName "1.0"
-
-        testInstrumentationRunner "com.moonlightbutterfly.cryptohub.CustomTestRunner"
     }
 
     buildFeatures {
@@ -115,7 +113,5 @@ dependencies {
     implementation("androidx.core:core-splashscreen:1.0.0")
     implementation 'androidx.hilt:hilt-work:1.0.0'
     implementation "androidx.work:work-runtime-ktx:2.8.0"
-    kapt "com.google.dagger:hilt-compiler:2.44"
-    kapt 'androidx.hilt:hilt-compiler:1.0.0'
     implementation 'com.jakewharton.timber:timber:5.0.1'
 }
