+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
     package="com.moonlightbutterfly.cryptohub">
 
     <uses-permission android:name="android.permission.INTERNET" />
@@ -14,16 +13,6 @@
         android:roundIcon="@mipmap/launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.AppSplash">
-
-        <meta-data
-            android:name="com.facebook.sdk.ClientToken"
-            android:value="@string/fb_app_client_token" />
-
-        <provider
-            android:name="androidx.startup.InitializationProvider"
-            android:authorities="${applicationId}.androidx-startup"
-            tools:node="remove">
-        </provider>
     </application>
 
 </manifest>
