--- app/src/main/res/values/strings.xml
@@ -1,64 +1,0 @@
-<resources>
-    <string name="app_name">CryptoHub</string>
-    <string name="crypto_asset_item_list_image_description">Crypto asset logo description</string>
-    <string name="crypto_assets_list">Crypto assets list</string>
-    <string name="settings">Settings</string>
-    <string name="night_mode">Night mode</string>
-    <string name="device_specific">Device specific</string>
-    <string name="statistics">Statistics</string>
-    <string name="market_cap">Market Capitalization</string>
-    <string name="volume">Volume</string>
-    <string name="rank">Rank</string>
-    <string name="circulating_supply">Circulating Supply</string>
-    <string name="max_supply">Max Supply</string>
-    <string name="favourites">Favourites</string>
-    <string name="search">Search</string>
-    <string name="cancel">Cancel</string>
-    <string name="recently_viewed">Recently viewed</string>
-    <string name="delete">Delete</string>
-    <string name="refresh">Refresh</string>
-    <string name="sign_out">Sign out</string>
-    <string name="continue_without_sign_in">Continue without sign in</string>
-    <string name="or">or</string>
-    <string name="google_icon">Google icon</string>
-    <string name="sign_in_failed">Sign in failed</string>
-    <string name="loading_data_failed">Loading data failed (probably because of API limit)</string>
-    <string name="email">Email</string>
-    <string name="password">Password</string>
-    <string name="continue_sign_in">Continue</string>
-    <string name="phone_number">Phone number</string>
-    <string name="code">Code</string>
-    <string name="phone_sign_in_description">After signing you might receive an SMS message for verification. Standard rates apply.</string>
-    <string name="sign_in">Sign in</string>
-    <string name="code_description">Type here the code you received through SMS</string>
-    <string name="facebook_application_id">998474207469300</string>
-    <string name="facebook_login_protocol_scheme">fb998474207469300</string>
-    <string name="fb_app_client_token">f0a4d6359aeb3c2d52e242cb9495a2ef</string>
-    <string name="fields_not_empty">Fields cannot be empty</string>
-    <string name="field_not_empty">Field cannot be empty</string>
-    <string name="network_error">Network error occurred</string>
-    <string name="not_found_error">Not found error occurred</string>
-    <string name="access_denied_error">Access denied error occurred</string>
-    <string name="service_unavailable_error">Service unavailable error occurred</string>
-    <string name="unknown_error">Unknown error occurred</string>
-    <string name="billion_abbr">bln</string>
-    <string name="million_abbr">mln</string>
-    <string name="channel_name">Crypto info</string>
-    <string name="channel_description">Informations about cryptocurrencies</string>
-    <string name="notification_data">Current %s price is %.2f </string>
-    <string name="notification_interval_30_minutes">30 Minutes</string>
-    <string name="notification_interval_hour">1 hour</string>
-    <string name="notification_interval_2_hours">2 hours</string>
-    <string name="notification_interval_5_hours">5 hours</string>
-    <string name="notification_interval_day">Day</string>
-    <string name="notification_interval_week">Week</string>
-    <string name="select_time">Select time</string>
-    <string name="clear_time">Clear time</string>
-    <string name="selected">Selected:</string>
-    <string name="save">Save</string>
-    <string name="every">every</string>
-    <string name="disable">Disable</string>
-    <string name="saved_notifications">Saved notifications: </string>
-    <string name="post_notifications">Post notifications</string>
-    <string name="notifications">Notifications</string>
-</resources>
