+++ app/src/main/res/values/themes.xml
@@ -11,43 +11,9 @@
         <item name="windowSplashScreenAnimatedIcon">@drawable/launcher_foreground</item>
         <item name="windowSplashScreenAnimationDuration">1000</item>
         <item name="postSplashScreenTheme">@style/Theme.CryptoHub</item>
-
-        <!-- Status bar and Nav bar configs -->
         <item name="android:statusBarColor">@color/fui_transparent</item>
         <item name="android:navigationBarColor">@color/yellow</item>
         <item name="android:windowLightStatusBar">true</item>
 
-    </style>
-    <style name="Theme.CryptoHub.Day">
-        <item name="colorPrimary">@color/yellow</item>
-        <item name="colorPrimaryVariant">@color/yellow</item>
-        <item name="colorOnPrimary">@color/white</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/red</item>
-        <item name="colorSecondaryVariant">@color/red</item>
-        <item name="colorOnSecondary">@color/white</item>
-        <item name="android:windowBackground">@color/white</item>
-    </style>
-
-    <style name="Theme.CryptoHub.TimePicker">
-        <item name="colorPrimary">@color/yellow</item>
-        <item name="colorPrimaryVariant">@color/yellow</item>
-        <item name="colorOnPrimary">@color/white</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/yellow</item>
-        <item name="colorSecondaryVariant">@color/yellow</item>
-        <item name="colorOnSecondary">@color/white</item>
-        <item name="android:windowBackground">@color/white</item>
-    </style>
-
-    <style name="Theme.CryptoHub.Night">
-        <item name="colorPrimary">@color/yellow_dark</item>
-        <item name="colorPrimaryVariant">@color/yellow_dark</item>
-        <item name="colorOnPrimary">@color/black</item>
-        <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/red_dark</item>
-        <item name="colorSecondaryVariant">@color/red_dark</item>
-        <item name="colorOnSecondary">@color/black</item>
-        <item name="android:windowBackground">@color/black</item>
     </style>
 </resources>
