+++ core/src/main/java/com/moonlightbutterfly/cryptohub/Constants.kt
@@ -1,0 +1,7 @@
+package com.moonlightbutterfly.cryptohub
+
+class Constants {
+    companion object {
+        const val NOTIFICATION_CHANNEL_ID = "general"
+    }
+}
