--- core/src/main/java/com/moonlightbutterfly/cryptohub/Dupa.kt
@@ -1,8 +1,0 @@
-package com.moonlightbutterfly.cryptohub
-
-class Dupa {
-
-    companion object {
-        const val NOTIFICATION_CHANNEL_ID = "general"
-    }
-}
