+++ data/build.gradle
@@ -61,6 +61,8 @@ dependencies {
     implementation 'androidx.hilt:hilt-work:1.0.0'
     implementation "com.google.dagger:hilt-android:2.44"
     kapt 'com.google.dagger:hilt-android-compiler:2.44'
+    kapt "com.google.dagger:hilt-compiler:2.44"
+    kapt 'androidx.hilt:hilt-compiler:1.0.0'
 
     kapt "androidx.room:room-compiler:2.5.0"
     implementation "androidx.room:room-runtime:2.5.0"
