+++ presentation/build.gradle
@@ -8,15 +8,15 @@ plugins {
 }
 
 android {
-    namespace 'com.moonlightbutterfly.presentation'
+    namespace 'com.moonlightbutterfly.cryptohub.presentation'
     compileSdk 33
 
     defaultConfig {
         minSdk 23
         targetSdk 33
 
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
         consumerProguardFiles "consumer-rules.pro"
+        testInstrumentationRunner "com.moonlightbutterfly.cryptohub.CustomTestRunner"
     }
 
     buildTypes {
