+++ presentation/src/main/java/com/moonlightbutterfly/cryptohub/panel/CryptoassetPanelComposables.kt
@@ -62,9 +62,9 @@ import com.moonlightbutterfly.cryptohub.models.CryptoAssetMarketInfo
 import com.moonlightbutterfly.cryptohub.models.NotificationConfiguration
 import com.moonlightbutterfly.cryptohub.models.NotificationInterval
 import com.moonlightbutterfly.cryptohub.models.NotificationTime
+import com.moonlightbutterfly.cryptohub.presentation.R
 import com.moonlightbutterfly.cryptohub.utils.round
 import com.moonlightbutterfly.cryptohub.utils.toStringAbbr
-import com.moonlightbutterfly.presentation.R
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.util.Calendar
