+++ presentation/src/main/java/com/moonlightbutterfly/cryptohub/signin/SignInComposables.kt
@@ -26,7 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.moonlightbutterfly.cryptohub.core.ErrorHandler
-import com.moonlightbutterfly.presentation.R
+import com.moonlightbutterfly.cryptohub.presentation.R
 
 val SIGN_IN_CONTENT_UNDER_LOGO_HEIGHT = 400.dp
